Updated the Main Body HTML
Added new div's to the main article section of the body to make the CSS work correctly. The new images I emailed over will just need to be added to the contents folder and the old ones removed. I changed the names so they are the same as the ones that need to be updated.